Fix compiling CellBroadcastReceiver with use_resource_processor: true
CellBroadcastReceiver fails to compile when use_resource_processor: true
is enabled by default because it compiles the same sources multiple times,
in CellBroadcastReceiver, CellBroadcastReceiverPlatform and
CellBroadcastReceiverOemUnitTests, each with a different manifest package
name.  When use_resource_processor: true is enabled the resources compiled
in the android_app are only available under the package name in the
manifest, so any package used in the sources will only work in one of
the android_app modules.  Move the compilation of the sources and resources
into a single android_library with the correct package name.  This also
avoids compiling the same sources multiple times.

Bug: 294256649
Test: m javac-check -k
Change-Id: I7342b98301f5be284a57d77515cfb3e8587f5d3d