Set TTS Field as mockTts during CellBroadcastAlertAudioTest
When unitest is run, CellBroadcastAlertAudio's onStartCommand is not run
on main thread while TextToSpeech called onInit() callback on main thread.
Because it's multi-thread, onInit callback can be called before TextToSpeech instance creation is completed.
Then, it makes nullpointerexception on mTts.
Set TTS field as mockTts after CellBroadcastAlertAudio instance is created and
before onStartCommand is called for preventing this exception.

Bug: 332807462
Test: atest com.android.cellbroadcastreceiver.unit.CellBroadcastAlertAudioTest --iteration 1000
Change-Id: I1adfb24575a30648112839f596fdd12fa794568e